typos

+++ FlowNet2_src/models/components/ops/resample2d/make.sh
@@ -4,13 +4,10 @@ TORCH=$($PYTHON -c "import os; import torch; print(os.path.dirname(torch.__file_
 
 cd src
 echo "Compiling resample2d kernels by nvcc..."
-if [ -f Resample2d_kernel.o ]; then
-    rm Resample2d_kernel.o
-fi
-if [ -d ../_ext ]; then
-    rm -r ../_ext
-fi
-nvcc -c -o Resample2d_kernel.o Resample2d_kernel.cu -x cu -Xcompiler -fPIC -arch=sm_30 -I ${TORCH}/lib/include/TH -I ${TORCH}/lib/include/THC -I ${TORCH}/lib/include -std=c++11
+[ -f Resample2d_kernel.o ] && rm Resample2d_kernel.o
+[ -d ../_ext ] && rm -r ../_ext
+
+nvcc -c -o Resample2d_kernel.o Resample2d_kernel.cu -x cu -Xcompiler -fPIC -arch=sm_50 -I ${TORCH}/lib/include/TH -I ${TORCH}/lib/include/THC -I ${TORCH}/lib/include -std=c++11
 
 cd ../
 $PYTHON build.py
